reviews вынесено в отдельный компонент
