+++ project/src/components/tabs/tabs.jsx
@@ -2,37 +2,19 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import filmProp from '../film/filmProp';
 import PropTypes from 'prop-types';
-import Review from '../review/review';
 import reviews from '../../mocks/reviews';
 import { TabNames } from '../../const';
 import Details from '../details/details';
+import Reviews from '../reviews/reviews';
 
 function FilmInfoData(film, tabName) {
   const filmReviews = reviews.filter((review)=>review.filmId===film.id);
-  const filmReviewsLeft = filmReviews.slice(0,filmReviews.length/2+1);
-  const filmReviewsRigth = filmReviews.slice(filmReviews.length/2+1, filmReviews.length);
+
   switch (tabName) {
     case TabNames.DETAILS:
       return <Details film={film}/>;
     case TabNames.REVIEWS:
-      return (
-        <div className="film-card__reviews film-card__row">
-          <div className="film-card__reviews-col">
-            {filmReviewsLeft.map((review, id) => {
-              const keyIndex = film.id + id + review.date;
-              return (
-                <Review key={keyIndex} review={review}/>
-              );})}
-          </div>
-          <div className="film-card__reviews-col">
-            {filmReviewsRigth.map((review, id) => {
-              const keyIndex = film.id + id + review.date;
-              return (
-                <Review key={keyIndex} review={review}/>
-              );})}
-          </div>
-        </div>
-      );
+      return <Reviews filmReviews={filmReviews}/>;
     default:
       return (
         <React.Fragment>
